Use Java 5

+++ key.core.symbolic_execution/src/main/java/de/uka/ilkd/key/symbolic_execution/strategy/SymbolicExecutionBreakpointStopCondition.java
@@ -70,7 +70,7 @@ public class SymbolicExecutionBreakpointStopCondition extends
             goal, node, ruleApp, executedNumberOfSetNodes);
         SourceElement activeStatement = NodeInfo.computeActiveStatement(ruleApp);
         if (isBreakpointHit(activeStatement, ruleApp, proof, node)) {
-            setMaximalNumberOfSetNodesToExecutePerGoal(executedNumberOfSetNodes.intValue());
+            setMaximalNumberOfSetNodesToExecutePerGoal(executedNumberOfSetNodes);
         }
     }
